refactor : Class 구분
