+++ src/main/java/racingcar/CarNameDecision.java
@@ -1,0 +1,17 @@
+package racingcar;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CarNameDecision {
+    List<RacingCar> racingCars = new ArrayList<>();
+
+    public List<RacingCar> Decision(String[] carNames){
+        for (int i = 0; i < carNames.length; i++) {
+            RacingCar racingCar = new RacingCar(carNames[i]);
+            racingCars.add(racingCar);
+        }
+    return racingCars;
+    }
+
+}
